Define contrasts inside a pipeline

Short wrapper around the contrasts<- function to define contrasts inside a pipeline easily. The function takes a dataframe as input and returns a dataframe with updated contrasts, allowing to pipe it, which differs from the usual use of the contrasts<- function.

Usage

define_contrasts(df, cols, contrasts)

Arguments

df

A dataframe.

cols

A vector of column names (quoted or not), e.g., c(group, condition).

contrasts

A list of contrasts, each specified either:

as a vector, e.g., c(-1, 1) for a two-level factor

as a matrix, e.g., contr.sum(3) for a three-level factor

Value

A dataframe with the contrasts defined for the supplied columns.

See also

Examples

df <- data.frame(
  group = factor(rep(c("A", "B"), 3)),
  condition = factor(rep(c("X", "Y", "Z"), each = 2))
)

# The usual way
contrasts(df$group) <- contr.sum(2)
contrasts(df$condition) <- contr.sum(3)

# Result
contrasts(df$group)
#>   [,1]
#> A    1
#> B   -1
contrasts(df$condition)
#>   [,1] [,2]
#> X    1    0
#> Y    0    1
#> Z   -1   -1

# the pipeline way
df <- df |>
  define_contrasts(
    cols = c(group, condition),
    contrasts = list(contr.sum(2), contr.sum(3))
  )

# Same result
contrasts(df$group)
#>   [,1]
#> A    1
#> B   -1
contrasts(df$condition)
#>   [,1] [,2]
#> X    1    0
#> Y    0    1
#> Z   -1   -1